requirements.txt
torch
ultralytics
onnx
onnxruntime
opencv-python
numpy
Pillow
pycocotools
psutil
tqdm
pandas
matplotlib
